Refactor admin auth
