+++ app/Http/Controllers/AuthController.php
@@ -3,10 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Events\UserSignedUp;
-use App\Http\Requests\Auth\AdminLoginRequest;
 use App\Http\Requests\Auth\LoginRequest;
 use App\Http\Requests\Auth\SignupRequest;
-use App\Http\Resources\LoggedInAdminResource;
 use App\Http\Resources\LoggedInUserResource;
 use App\Services\AuthService;
 use Illuminate\Http\JsonResponse;
@@ -142,84 +140,6 @@ class AuthController extends Controller
     public function logout(Request $request): JsonResponse
     {
         $this->authService->logoutUser($request->user());
-
-        return Response::json(null, HttpResponse::HTTP_NO_CONTENT);
-    }
-
-    /**
-     * Login an admin.
-     *
-     * @param  AdminLoginRequest  $request
-     * @return JsonResponse
-     *
-     * @throws HttpException
-     * @throws NotFoundHttpException
-     */
-    #[OAT\Post(
-        tags: ['adminAuth'],
-        path: '/api/admin/login',
-        summary: 'Login an admin',
-        operationId: 'AuthController.adminLogin',
-        requestBody: new OAT\RequestBody(
-            required: true,
-            content: new OAT\JsonContent(ref: '#/components/schemas/AdminLoginRequest')
-
-        ),
-        responses: [
-            new OAT\Response(
-                response: HttpResponse::HTTP_OK,
-                description: 'Ok',
-                content: new OAT\JsonContent(ref: '#/components/schemas/LoggedInAdminResource')
-            ),
-            new OAT\Response(
-                response: HttpResponse::HTTP_UNPROCESSABLE_ENTITY,
-                description: 'Unprocessable entity',
-                content: new OAT\JsonContent(ref: '#/components/schemas/ValidationError')
-            ),
-            new OAT\Response(
-                response: HttpResponse::HTTP_UNAUTHORIZED,
-                description: 'Unauthorized',
-                content: new OAT\JsonContent(
-                    properties: [
-                        new OAT\Property(
-                            property: 'message',
-                            type: 'string',
-                            example: 'Invalid credentials.'
-                        ),
-                    ]
-                )
-            ),
-        ]
-    )]
-    public function adminLogin(AdminLoginRequest $request): JsonResponse
-    {
-        $admin = $this->authService->loginAdmin($request);
-
-        return Response::json(new LoggedInAdminResource($admin));
-    }
-
-    /**
-     * Logout an admin.
-     *
-     * @param  Request  $request
-     * @return JsonResponse
-     */
-    #[OAT\Post(
-        tags: ['adminAuth'],
-        path: '/api/admin/logout',
-        summary: 'Logout an admin',
-        operationId: 'AuthController.adminLogout',
-        security: [['BearerToken' => []]],
-        responses: [
-            new OAT\Response(
-                response: HttpResponse::HTTP_NO_CONTENT,
-                description: 'No content'
-            ),
-        ]
-    )]
-    public function adminLogout(Request $request): JsonResponse
-    {
-        $this->authService->logoutAdmin($request->user());
 
         return Response::json(null, HttpResponse::HTTP_NO_CONTENT);
     }
