+++ routes/api/admin/auth.php
@@ -1,0 +1,7 @@
+<?php
+
+use App\Http\Controllers\Admin\AuthController;
+use Illuminate\Support\Facades\Route;
+
+Route::post('/login', [AuthController::class, 'login']);
+Route::post('/logout', [AuthController::class, 'logout'])->middleware('auth:admin');
