+++ routes/api/auth.php
@@ -6,8 +6,3 @@ use Illuminate\Support\Facades\Route;
 Route::post('/signup', [AuthController::class, 'signup']);
 Route::post('/login', [AuthController::class, 'login']);
 Route::post('/logout', [AuthController::class, 'logout'])->middleware('auth:user');
-
-Route::group(['prefix' => 'admin'], function () {
-    Route::post('/login', [AuthController::class, 'adminLogin']);
-    Route::post('/logout', [AuthController::class, 'adminLogout'])->middleware('auth:admin');
-});
